Time: 4 ms (32.51%), Space: 8.6 MB (67.65%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -4,7 +4,7 @@ public:
         // Negative numbers are never palindrome
         if (x < 0) return false;
 
-        long reversed = 0;
+        long long reversed = 0;
         int original = x;
 
         while (x != 0) {
